Add boilerplate code
Adapt it from lms/repositories/knote/lab01 to be similar to the code
shown in the instructions.

In the current state, the app will not launch, because there are no npm
dependencies declared. So, add to the instructions something like:

"The first thing you have to do is to install the dependencies that were
discussed above for your app. You can do this by running:

npm install express pug path

Now you can start the app with `npm start`."