Replace StatefulSet with Deployment for MongoDB
When a StatefulSet is deleted, the associated PVCs are not deleted,
which means that also the PVs are not deleted. So when deleting the
MongoDB StatefulSet and then starting it up again, the old data is still
in the database, which might be confusing.

By using a Deployment instead of a StatefulSet, it is easier to start
with a clean slate.